% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templator.R
\name{HtmlInspector}
\alias{HtmlInspector}
\title{Create a html file based on a template and a dataframe of data to insert}
\usage{
HtmlInspector(template.path, output.path, df = NULL, singles = list(title =
  "inspect segments"))
}
\arguments{
\item{df}{dataframe; contains the data that will be inserted in a reapeater}

\item{template.file:}{string; The name of the file that will be used as the template,
which is basically a html file with some specially delimeted flags
showing where the data should be inserted.}

\item{output.fn:}{string; where the html file should be saved}

\item{singles:}{list; extra data that is not contained in the data.frame that will be inserted as singles (e.g. page title)}
}
\value{
string; The path where the html file was saved
}
\description{
Create a html file based on a template and a dataframe of data to insert
}

